backend/requirements.txt
Django==3.2.16
djangorestframework==3.12.4
djoser
django-filter
psycopg2-binary
pytest==6.2.4
pytest-pythonpath==0.7.3
pytest-django==4.4.0
Pillow==9.3.0
requests==2.26.0
gunicorn
